requirements.txt
PyYAML==5.3.1
systemd-python==234